Separate package for notifier implementations
