+++ internal/notifications/gotify/gotify_test.go
@@ -1,4 +1,4 @@
-package notifications
+package gotify
 
 import (
 	"encoding/json"
